taup: Fix branch splitting.
Ray parameters are ordered largest to smallest, so the reversed
condition in TauModel.splitBranch means splitting *never* occurred.
Various other code paths were buggy, but never exercised. The
optimization I applied in TauBranch.difference did not handle the fact
that the difference should be between a split branch and a non-split
branch. Finally, TauBranch.insert was calculating branch for the wrong
set of layers.